Move session method to extension utils
