+++ product/web3modal/src/main/kotlin/com/walletconnect/web3/modal/domain/SessionRepository.kt
@@ -29,9 +29,9 @@ internal class SessionRepository(
             preferences[SELECTED_CHAIN]
         }
 
-    suspend fun getSelectedChain() = runBlocking { selectedChain.first() }
+    suspend fun getSelectedChain(): String? = runBlocking { selectedChain.first() }
 
-    suspend fun getSessionTopic() = runBlocking { sessionTopic.first() }
+    suspend fun getSessionTopic(): String? = runBlocking { sessionTopic.first() }
 
     suspend fun saveSessionTopic(topic: String) {
         context.sessionStore.edit { store ->
